Changed cmd_help to be more testable
No longer creating the help table every time - this is a constant, so I
am building it at module scope (should make things more snappy).

No longer creating the table in the tests.  It is just a string constant
in the test module.

added 2 tests, and modified one to test these changes.